Fixed all warnings

+++ src/main.c
@@ -13,6 +13,7 @@
 #include <csr.h>
 #include <trap.h>
 #include <block.h>
+#include <rng.h>
 
 // Global MMU table for the kernel. This is used throughout
 // the kernel.
@@ -120,7 +121,7 @@ static void init_systems(void)
 
     uint8_t buffer[16] = {0};
     debugf("RNG State Before:");
-    for (int i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
+    for (uint64_t i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
         debugf(" %d ", buffer[i]);
     }
     debugf("\n");
@@ -128,23 +129,23 @@ static void init_systems(void)
     debugf("RNG init done; about to fill\n");
     rng_fill(buffer, 16);
     debugf("RNG State After:");
-    for (int i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
+    for (uint64_t i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
         debugf(" %d ", buffer[i]);
     }
     
     rng_fill(buffer, 16);
-    for (int i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
+    for (uint64_t i=0; i<sizeof(buffer)/sizeof(buffer[0]); i++) {
         debugf(" %d ", buffer[i]);
     }
 
 
     uint8_t sector[512];
-    for (int i=0; i<sizeof(sector)/sizeof(sector[i]); i++) {
+    for (uint64_t i=0; i<sizeof(sector)/sizeof(sector[i]); i++) {
         sector[i] = i;
     }
     block_device_write_sector(0, sector);
 
-    for (int i=0; i<sizeof(sector)/sizeof(sector[i])/8; i+=8) {
+    for (uint64_t i=0; i<sizeof(sector)/sizeof(sector[i])/8; i+=8) {
         debugf("%d %d %d %d %d %d %d %d\n",
             sector[i], sector[i+1], sector[i+2], sector[i+3],
             sector[i+4], sector[i+5], sector[i+6], sector[i+7]);
@@ -152,7 +153,7 @@ static void init_systems(void)
 
     // block_device_write_sector(0, sector);
 
-    for (int i=0; i<sizeof(sector)/sizeof(sector[i]); i++) {
+    for (uint64_t i=0; i<sizeof(sector)/sizeof(sector[i]); i++) {
         sector[i] = 0;
     }
     // block_device_read_sector(0, sector);
